api: initial commit (rfc)
This commit adds an absolute minimal implementation of a driver,
capable only of handshaking with Scylla and sending queries,
not even waiting for responses. On top of it, we can start discussing
how the final api is going to look like.
Comments welcome!
This commit comes with a test which requires a running local Scylla
instance.

Co-authored-by: Piotr Sarna <sarna@scylladb.com>