prepared: add computing the partition key
In order to route the statement in a token-aware way,
the partition key needs to first be extracted from the prepared
statement and its bound values.
The computations were verified semi-manually by comparing the output
of Scylla's SELECT token(a,b,c) FROM ks.t, and 64 least significant
bits of murmur3 output.