Implement paging
This PR implements paging for both prepared and unprepared queries. A
new variant of `query` and `execute` methods is introduced -
`query_iter` and `execute_iter`, which returns a RowIterator.

The user operates on a RowIterator, which returns consecutive rows. When
a RowIterator is created, a worker task is spawned which fetches
consecutive pages and passes them to the RowIterator through a channel.

Fixes #10