types.rs: abstract away length serialization and raw bytes read
Adds new functions for reading and writing lengths:

  read_int_length
  read_short_length
  write_int_length
  write_short_length

Those functions check that the length being read/written is correct.

Additionally, a function for reading raw bytes is added which checks
that the slice has enough bytes and returns an error if it does not.
This is better than just trying to take a subslice, because trying to
take a subslice which is out of bounds causes a panic.

All serialization functions are updated to use the new helper functions
so that we don't miss some checks.