get rid of close()
futures_util::future::RemoteHandle is now used to schedule Connection's
workers for destruction by tokio's executor when Connection is dropped.
Yay for RAII.

There is no need for an explicit "wait for stuff to finish" method,
really. If the user wants there to be no in-flight requests when
they drop the Connection or Session, just wait for the requests to
finish before dropping.