switch to using "snap" crate instead of "snappy"
Currently, we are using the "snappy" crate for compression/decompression
of frames. This crate is just a binding to the C++ snappy library, and
therefore requires it to already be installed when building the crate.

This commit removes "snappy" crate as a dependency; now, the "snap"
crate is used, which is a pure Rust implementation of the snappy
algorithm with performance comparable to the C++ library. Now, the C++
library doesn't have to be installed when compiling the driver.

This change makes the driver compile and pass all units test on my
Windows installation.

Fixes #129