Convert HashMap to BTreeMap for CQLTypes
It is required that all CQLTypes have Hash trait implemented since every CQLType could be a key in a Map. HashMap doesn't provide default implementation for Hash and is unordered so two maps with the same values and keys could return two hash values which could lead to future bugs.