SessionBuilder uses openssl's SslContext for TLS support
User should be able to decide what kind of certificates, files and verification modes they
are using, everything can be configured inside of SslContext struct that we're going to use
to create an SSL connection.
Tests were adjusted.
Test for setter of ssl_context were ommited due to SslContext not implementing PartialEq trait.
Created example of tls in use.