frame: introduce SerializedRequest
Introduces SerializedRequest, which is a wrapper over a vector
containing a fully serialized request frame. Keeping serialized header
and body in a contiguous vector makes it possible to send them using a
single `write_all` request, which can translate to a single syscall.

Previously, header and body were kept separately and had to be written
using two calls which can result in two packets being generated, which
is wasteful. Moreover, with Nagle's algorithm enabled on the client side
and delayed ACK enabled on the Scylla side, this would result in
latencies being extremely bad if not enough workload was generated to
saturate connections.

Because serialization is done in a separate task from stream ID
assignment, the SerializedRequest allows to set the stream ID directly
in the serialized frame.