Add add_metrics method to SessionConfig
This change fixes the problem of not being able to initialize
`PercentileSpeculativeExecutionPolicy` with `MetricsView` when creating
session. It is possible now to create metrics first, feed them to
`PercentileSpeculativeExecutionPolicy` and then create session with
speculative execution policy set to the one previously mentioned policy.

This commit also removes `MatricsView` which was used to restrict
visibility of `Metrics` public methods (I had changed all of `Metrics`
inc* methods which should not be used by end user to `pub(crate)`, so
`MetricsView` was redundant).