connection: propagate authentication failure error to user
Currently, during authentication, the "ERROR" frame is treated as
unexpected and an error with the following message is returned to the
user:

    Unexpected response to Authenticate Response message

This commit causes the error from the "ERROR" frame to be propagated to
the user in case of authentication failure. For example, if bad
username/password combination is provided, Scylla will return the
following message:

    Username and/or password are incorrect

which is much more helpful.

Fixes #257