routing: introduce Sharder
ShardInfo represents information related to sharding which was parsed
from OPTIONS response sent by the database. It contains the id of the
connection's shard, total number of shards and the msb_ignore parameter
used for calculating shard based on token.

ShardInfo has multiple methods used for calculations related to
sharding, e.g. to calculate the shard for a token or iterate over source
ports which can be used to connect to shard-aware port for a given
shard.

However, none of those methods use the first field of ShardInfo which
represents the id of the connection's shard. If a method needs to
calculate something for a particular shard id, the id is passed as an
argument.

This commit moves most of ShardInfo's methods to a new struct, Sharder.
The new struct is identical to the first one, but does not have the
"connection's shard id" parameter.