node: make Node constructor async
The upcoming connection pool (which will be introduced in the next
commit of this PR) will attempt to synchronously fill itself on its
construction. Because the pool will be constructed when the Node is
constructed, the Node constructor have to be marked as async. In order
to reduce noise, the Node constructor is marked as async in this,
separate commit.