docs: remove stray hashes from rust-driver.docs.scylladb.com
Sphinx, the framework used for publishing docs on
rust-driver.docs.scylladb.com, can't handle mdbook's comments
starting with '#', which results in broken docs pages.
A simple heuristics for removing incorrectly generated snippets
is added to the deploy script - it will work as long as sphinx
generates HTML the same way, so it may need to be revamped
in the future.