topology: always update known_peers on refresh (#270)
Currently, the TopologyReader is responsible for maintaining the control
connection. When the control connection breaks, it attempts to
re-establish by trying nodes from the `known_peers` list in random
order.

The `known_peers` list is modified only in two situations:

- It is initialized from the known nodes list from the session's
  configuration,
- When topology refresh is requested, current control connection breaks
  and is successfully reconnected, the `known_peers` list is
  reinitialized from fetched topology info.

However, this is not enough. If only known node is provided in the
session configuration and the node is removed from the cluster, the
TopologyReader won't be able to reconnect the control connection because
the only node from the `known_peers` list has just been removed.

The solution to this is to update the `known_peers` list on each
topology refresh. This way, the `known_peers` list will be up-to-date
and the TopologyReader will consider all nodes in the cluster when
re-establishing the control connection.

Fixes #269