frame: fix serialization of bytes length in AUTH_RESPONSE
The body of AUTH_RESPONSE frame is a single [bytes] value - it should
start with a i32 length and then the contents should follow. The length
was incorrectly encoded as u8. While Scylla didn't complain, Cassandra
returned an `IndexOutOfBoundsException` because it interpreted the
beginning of the [bytes] contents as garbage length.

Now, the length is properly encoded as i32, and serialization of
username and password is slightly optimized.