frame: change `CqlValue`'s UDT representation
`CqlValue` has a variant called `UserDefinedType`, that is responsible
for storing UDT responses. It stored information about UDT fields using
`BTreeMap<String, Option<CqlValue>>` - key was a field's name, value was
a field's value. Because of using `BTreeMap`, an order of fields defined
by the type was lost. Hence, the original order of the fields could not
be restored if someone's only data was a `CqlValue` (this made `Value`
for `CqlValue::UserDefinedType` impl impossible).
This commit swaps `CqlValue::UserDefinedType`'s `BTreeMap` for `Vec` of
key-value tuples. It allows implementing `Value` for
`CqlValue::UserDefinedType` and speeds up UDP deserialization (by
removing the need to construct temporary `BTreeMap`).