topology: drop the serde_json dependency (#328)
At the time when the logic responsible for querying keyspace metadata
was implemented, there was no support for map datatypes, which was
needed to parse the `replication` parameters of a keyspace. To work
around that, the `toJson` function was used so that Scylla returns the
parameters encoded as a JSON string. The serde_json library was added to
dependencies solely for the purpose of parsing the encoded json map.

Support for direct deserialization of map types has been merged some
time ago, so there is no need to involve JSON as an intermediate step.

This commit removes the logic responsible for parsing the `replication`
map from JSON representation and now the map is parsed directly from its
CQL representation. Additionally, the serde_json crate is removed from
dependencies.