connection: buffer writes if tasks are available in the channel
This commit adds buffering capability to the writer.
It works as follows - once a task was properly received
from the channel, an attempt is made to try receiving more data
from the same channel in a non-blocking way. If data is available,
it will be written immediately into the buffered writer (BufWriter).
Flushing the buffer explicitly is only performed after it's no
longer possible to fetch data from the channel without waiting.

NOTICE: write buffering is done via a BufWriter class
due to the fact that Tokio does not yet expose write_all_vectored,
and its implementation of write_vectored does not fit our needs,
because it basically just sends one buffer at a time anyway.
Once write_all_vectored is solved, write buffering can be
simplified, and memory copying can be elided.