topology: ignore non-standard tokens
When fetching topology, our previous reaction to a non-standard
token was to fail the driver, which is harsh - especially on
managed solutions like Amazon Keyspaces, which does not allow
changing the partitioning scheme.
Instead, a dummy token is used, a warning is printed and a FIXME
is added to address the problem more thoroughly in the future.

Fixes #350