docs: Replace double trailing space with backslash
In markdown to make a new line we can write:
First line<space><space>
Second line.

Without the double space both lines would be rendered in the same line.
It acts like <br> in HTML
We use this in our documentation.
However there is also another way to specify a newline:
First line\
Second line

In my opinion the second way is better:
* It is more visible when writing the markdown
* Contributors remove trailing spaces, because they think it is a mistake
* Some editors automatically remove trailing spaces

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>