transport: break connections with large orphaned stream id count
The age of orphaned stream ids is tracked. Once a connection has too
many old ones, it breaks, and a new connection should be established
instead. Making decisions based on the number of old orphans rather than
the number of all orphans, prevents braking healthy connections in which
a user dropped a lot of queries (e.g. made a lot of queries using
speculetive execution mechanism).