topology: make control connection more persistent
After this patch, trying to read topology metadata does
not ever end on its own - instead, the function retries
in an infinite loop. It's the caller's responsibility to
eventually stop waiting and abort - right now it is done
after 60 seconds (our hardcoded refresh period), but we
should also consider making it configurable.

Fixes #398