CI: don't use a prebuilt container in the "Book" workflow (#403)
The "Book" workflow uses a prebuilt Docker image "scylla-mdbook" which,
in addition to the usual Rust compilation tools, contains a `cargo
mdbook` command installed. The problem with this approach is that the
container should be rebuilt regularly in order to update the Rust
version within - however, nobody remembers to do that and after 10
months the compiler became so outdated that it no longer recognises some
of the newly introduced changes to the language.

This commit changes the "Book" workflow so that it bases on
`ubuntu-latest` which is updated regularly, and the `cargo mdbook` tool
is installed as the first step. Although this is extra work on each CI
run compared to the previous approach, it requires less maintenance and
is less likely to break in the future.