transport: retry_policy: update retry conditions
Adds a RetryNextNode decision for the QueryError::UnableToAllocStreamId.
This error happens when a request cannot be send to Scylla because the
chosen connection is overloaded. It is desired for a retry session to
retry on the next node when seeing this error as the next node's
connections may not be overloaded.