frame: allow using Time struct as CqlValue
It's particularly useful in IntoUserType and FromUserType macros
to be able to express that a column has CQL Time type.
For that we need a wrapper, and a perfect candidate for that
is scylla::frame::value::Time, which is already a thin wrapper over
chrono::Duration. This patch simply implements FromCqlVal for this
struct, which will allow using it for serialization and deserialization.

Fixes #421