iterator: provide default page size when querying with _iter
Both `Query` and `PreparedStatement` have page size set to `None` as
default, which means unlimited page size. This is a problem for
`query_iter` and `execute_iter` because using them with such queries
causes everything to be fetched in one page, despite them being meant
to fetch data page-by-page.

We can't really change the default page size for `Query`
and `PreparedStatement` because it also affects
`Session::{query,execute}` and this could break existing code.

In order to work around the problem we just set the page size to
a default value at the beginning of `query_iter` and `execute_iter`.

The default is 5000 rows, just like the default page size in gocql and
the Java driver.

Fixes #424