examples: add example tower::Service trait implementation
This commit provides a simple example on how to wrap a Scylla session
into a tower::Service, which gives such a session all the benefits
of tower - composability, universal API, etc.
The example provides a very trivial service which is only capable
of sending unprepared, parameterless queries, but it's educational
nonetheless. By implementing a service on top of CachingSession
wrapped with a list of parameters, a full-fledged service can be
easily implemented.

Refs #425