cluster: use dummy metadata when initial fetch fails (#440)
Currently, the driver fetches topology and schema information (we call
it metadata) on Session creation and fails if an error occurs during
fetching/parsing of that data - despite the fact that the metadata is
not vital for the driver to work (although some information is needed
for better performance, e.g. topology info and keyspace replication
strategies).

While, in ideal world, we would like the metadata fetch to always
succeed, it can fail either due to our fault (e.g. bugs like #437) or
reasons beyond our control (connection errors, non-backwards compatible
changes in system tables). Therefore, we would like for the driver to
work even when no metadata is available.

This commit modifies the initial metadata fetch code so that, in case it
fails, a dummy metadata is used instead.

Fixes: #438