Optimize PreparedStatement::compute_partition_key (#436)
In PreparedMetadata initialization, PK indexes are associated to their
sequence number, and then sorted by the index value. It allows iterating on
bounded_values in order, because indexes can then be reordered thanks to
sequence number.

Also, buf size is computed during iteration in order to be reserved before
writing to it, preventing buffer reallocation at each write.

Last but not least, an array is used instead of a vector where the size of
pk_indexes is not too big, saving the allocation cost.

Quick benchmark using `complex_pk` table of `test_prepared_statement` shows a
performance improvement from 200ns to at 130ns.