CI: check compilation with MSRV
As we are about to declare 1.59 as our minimum supported rust version,
this commit introduces a CI check that will help us enforce it. It runs
the `cargo check` command using the MSRV toolchain and verifies that it
passes with and without features enabled.

The check was inspired by tokio's CI configuration.