transport: topology: add up/down info to Peer
For node up/down markers to work reliably, a source of truth other than
CQL events is needed. That's because the CQL events delivery is prone to
connection failure. By quering `system.cluster_status` table from time
to time, we are able to get the true information about the status of
nodes in the cluster.

This change will enable us to correct the up/down markers because every
`Node` is created (and also can be updated) based on information from a
corresponding `Peer`.