Allow to access PreparedMetadata
After preparing a statement the database
returns a bunch of metadata information
that is kept in PreparedStatement.
This data isn't normally used by
users of the driver so it's private.

While it might not be useful to regular
users I was just debugging a problem with Scylla
and needed to access it.
Let's make it accessible from outside
to make life easier for database developers
who use this driver.

Signed-off-by: cvybhu <jan.ciolek@scylladb.com>