cluster/topology: use dummy metadata only on first fetch error, not connection error
Now, `fetch_metadata` differentiates between a failure to establish a
connection and failure to fetch metadata. In case of the second, if the
`initial` flag is true, it ignores the error and returns dummy metadata.

The dummy metadata replacement logic in `Cluster::new` is superseded by
`fetch_metadata` and was removed.