session_test: add test for failed session connect
Adds a test which verifies that creating a session fails when the
control connection fails to be established. It spawns a micro-server
which accepts and then immediately closes connections, and then verifies
that creating a session and trying to connect to this server fails with
an error.

At the moment, it reports the following error:

Connection error (it was expected): IoError(Custom { kind: Other, error: "No connections in the pool; last connection failed with: Invalid message: Frame error: early eof" })