session: route prepared queries without pk index like unprepared queries
Not all prepared statements can be routed in a token-aware manner.
Sometimes the token cannot be computed by the driver based on the bound
query values - if this is the case, PREPARE will return metadata with
empty pk_index (which tells the driver how to compute the token).
However, the driver interprets empty pk_index as valid information and
computes a dummy token which is then used to route the statement to the
same node/shard every time.

This commit changes the execute logic so that it detects when the
prepared statement has empty pk_index and routes it in the same way as
unprepared queries.