session: automatically wait for schema agreement
Now, the session automatically waits for schema agreement after
executing a statement which modifies the schema. The rationale for this
change is that, previously, it was the user's responsibility to wait for
schema agreement and it was easy to forget about this check.

The automatic wait for schema agreement is never a bug, although it can
slow down some programs which do a lot of schema changes in a short
succession. Those applications can opt-out of this feature by disabling
the feature in the configuration and can still wait for schema agreement
manually.

The session will wait for schema agreement for a configured amount of
time, and then - if the agreement is not reached - it will return
a TimeoutError.

Fixes: #459