Optional address translation added
Added AddressTranslator trait and an ability to provide a struct
implementing it to the session builder. When done so, every address
fetched as cluster metadata will be translated using the rules defined
by the trait implementor.
Also, removed the special treatment of the node that the driver connects
to first: now, the address received from system tables replaces the
address used to connect to that node in the first place. The reason for
that was incompatibility with address translation: known_peers vector
would be holding a mix of translated and untranslated addresses.