session_test: Test repreparation in CachingSession::execute()
Add a test which checks what happens when the prepared
statement cache	gets cleared and the PreparedStatement
needs to be prepared again.

CachingSession::execute() should handle this gracefully
and automatically reprepare the	statement.

In case	there was a schema change and the statement
is no longer valid the code shouldn't panic -
it should return an error.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>