caching_session: execute: Take ValueList by value
CachingSession::execute takes the value list
argument as:
values: &impl ValueList,

This is an unnecessary limitation. Session has:
values: impl ValueList,
instead.

This forces the user to pass values as a reference.
So instead of writing:
session.execute(..., (1, 2, 3))
They have to write:
session.execute(...., &(1, 2, 3))

There is a generic impl that makes every refernce
to ValueList also implement ValueList:
impl<T: ValueList> ValueList for &T {
     ...
}

This means that this is not a breaking change.
All existing code will work without any modifications
(the tests didn't need to be modified).
It just enables the user to write cleaner code.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>