transport: locator: add PrecomputedReplicas struct
`PrecomputedReplicas` is responsible for precomputing all possible
replica lists for given replication strategies. Having the replicas
precomputed allows to avoid costly calculations each time a request is
sent. To get a replica list for a given token the driver can simply do a
lookup in the precomputed data.

Due to the fact that:
    Replica lists for replication factor `n` are prefixes of replica
    lists for replication factor `n + 1` ony if requested `replication
    factor` is <= `rack count` (when considering Simple Strategy, one
    can assume `rack count` to be equal to the total number of nodes in
    the cluster).

We were able to only compute replica lists for maximal replication
factor found in replication strategies - which saves memory and CPU
resources.

Co-authored-by: Michał Sala <michal.sala@scylladb.com>