topology: use rpc_address instead of peer when fetching addrs
When fetching node addresses from system tables, `peer` column
is not the right choice - the address available for clients
is generally stored in `rpc_address` column.
Technically, a full solution should also fetch `broadcast_rpc_address`
and use it if `rpc_address` is incorrect or set to 0.0.0.0,
but this quick fix is enough to work for most configurations.