Fixed bug: no retries performed on Error response
Because of bug in `Session::execute_query()`, no retries were
performed if Scylla returned Response::Error. This was because such
response was considered to be an Ok() variant and not Err() one.
This patch removed Response::Error enum variant completely and rewrites
the logic in a way such that the information about such errors
is passed as Err(QueryError) variant.