session: Use NonErrorQueryResponse in run_query
run_query takes as a paramter a closure
used to run a single query ona connection.
Result of such a query is Result<ResT, QueryResult>.

In Case of Result::Err the retry policy decides
whether to retry the query based on the error type.

There was a bug - in a few cases ResT was
of type QueryResponse, which has an Error variant.
Such errors weren't detected by RetryPolicy
and were never retried.

This commit fixes the bug by making the
closure return NonErrorQueryResponse.
All errors will be collected in
a single QueryResult and detected
by RetryPolicy.

Fixes: #501

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>