proxy: added manual sanity check method
As errors that occur in proxy are asynchronous, they must wait for a
proper moment to be examined by the caller. However, in some tests using
proxy the results may be unpleasantly unclear because we only see the
failing assertion and not the error that was the primary cause. Hence,
a new method `sanity_check()` was added to RunningProxy that enables
propagating errors from the proxy during its runtime, not only after it
finishes.