iterator: Add history collecting in iterators
When history_listener in StatementConfig
is set to Some, iterators should log
execution history.

Implement this.
This covers Session::query_iter,execute_iter

Each page query is trated as a separate Query.

Iterators don't use speculative execution
so it isn't logged anywhere.

The implementation is a bit funky because
iterators have a bit of a strange flow.
query_pages fetches pages until there is
an error, so it can span multiple queries
or none at all.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>