docs: Preprocess docs for multiversion target
sphinx-multiversion is now called with an
argument that makes it preprocess the source
directory. This happens in a temporary directory
so the original files are not modifed in any way.

The sad thing is that it doesn't really work
on this branch because sphinx-multiversion
checks out to main where the preprocessing script
doesn't exist :/
A solution would be to merge the prepocessing script
first and then this commit.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>