test: Move unique_name() to utils::test_utils
unique_name() is a function defined in session_test.rs.
It's used to generate an unique keyspace name during testing.

Many other modules also used this function by calling
crate::transport::session_test::unique_name().

This is a bit ugly, it would be better to define
this function in a global place which all tests
could access. Additionally tests that aren't
a submodule of the transport module can't access
it because the module is private.

It has been now moved to crate::utils::test_utils
and renamed to unique_keyspace_name().

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>