session_test: add test for the rate limit exceeded
Adds a test which checks that the driver properly negotiates support for
the RateLimitExceeded exception and receives it from Scylla when it is
triggered.

Because this feature is Scylla-specific and will only be supported in an
upcoming version of Scylla, the test first checks that the cluster
supports the feature - if not, it is skipped.