utils: add ArmableTimer
The ArmableTimer is a future which is also a resettable timer. It starts
in an "unarmed" state. If armed, it resolves at the configured point of
time. After successfully polling the timer after it resolves, it becomes
unarmed and must be reset again.

It is possible to concurrently await the timer and (re)-arm. Awaiting on
an unarmed timer does not resolve unless the timer is armed and the
deadline is reached. Re-arming a timer while awaiting on it will just
change the point of time when the awaiter resolves.

The ArmableTimer will be used as a component of the new, scalable
reimplementation of timeouts.