connection: add DeadlineTracker
Adds a module to the connection router which is responsible for tracking
request deadlines and sending timeouts errors for their handlers when
their deadlines pass.

The main purpose of tracking deadlines on the router rather than via
tokio::time::timeout on the tasks that wait for response is to avoid
scalability issues. The current implementation of Tokio's `timeout`
registers a timer entry in an intrusive list in the current thread's
timer driver. If the inner future finishes before the timeout, the
`timeout` future needs to unregister the timer entry. Both registration
and unregistration requires acquiring a per-timer driver mutex, and we
observed that under very large throughput and core count acquiring this
mutex takes a lot of the total execution time (note that tokio worker
threads can steal tasks from each other, so cancelling a timer might
involve acquiring a mutex of a different thread's timer driver).

The new implementation reduces the number of times timers are registered
and unregistered. The deadline tracker module keeps a single timer and
reschedules it infrequently, i.e. roughly with a period of timeout minus
latency (the assumption is that the timeouts are rare and are longer
than the average latency).