connection: introduce deadline-aware versions of the Connections' methods
Because we are moving responsibility for timing out the request from the
task waiting for the request to the connection router, the request
deadline needs to be propagated via Connection's methods such as query,
execute or batch. In order not to break the existing Connection's
interface, this commit introduces new methods which take an optional
deadline parameter and use them to reimplement the existing methods. The
new methods have pub(crate) visibility in order not to pollute the
public interface with even more versions of query/execute/batch methods.