iterator: refactor parameters of new_for_query into a struct
The `RowIterator::new_for_query` has a lot of parameters, and we will be
adding more of them. Most of them are shared with
`new_for_prepared_statement` which takes those parameters as a struct.
This commit refactors the code so that both of these methods take a
common struct.