topology: filter hosts for the control connection
Now, the control connection is aware of the host filter and carefully
tries to omit filtered out nodes.

If the initial nodes list contained a node that is not accepted by the
host filter and a control connection has been established to it, a
warning will be printed and the driver will try to re-establish it to
one of the accepted nodes.

Additionally, in case of a mistake where an incorrect host filter is
provided and all nodes are filtered out, a message is printed which
points out the problem.