prepared: added is_confirmed_lwt field and getter
This will enable taking into consideration whether the statement is
a LWT. If so, its queries will be routed to the replicas in the fixed
order (as it speeds LWT execution up significantly).
The extension of PreparedStatement was also reflected in CachingSession,
by storing is_confirmed_lwt value in RawPreparedStatementData, too.