make: wait for the cluster to be ready before running tests
Running `make up` and `make test` in a consecutive order could cause
tests to fail due to the fact that ScyllaDB cluster did not have the
time to finish its initiation.

Added `wait-for-cluster` make target and `scripts/wait_for_cluster.sh`
script allow to wait for the possibility to connect to the native
transport port of each node.

`up` and `wait-for-cluster` targets were specified as dependencies of
the `test` make target. Thanks to this change, `make test` will wait for
the cluster to become ready before running test. The `make up` command
is also no longer needed before invocation of `make test`.