test: make dockerized tests use the same docker-compose.yml as normal ones
Tests in Docker (scripts/test_in_docker/) were modified to use the same
docker-compose.yml as the CI and Makefile. They were also moved to
test/dockerized/ directory, and made to be run by a `dockerized-test`
make target. I chose to drop the test_in_docker.ps1 script that was
written to make it easier to run tests on Windows. Running tests in a
Docker container on Windows changes the build target to the Linux one -
Windows-specific code paths are not tested in that way. I don't see any
benefit of maintaining this script, especially given that Windows users
can use WSL to run tests.
As for now, I think that test/dockerized makes sense only on macOS.