connection: clearly mark db-side warnings as such
The CQL protocol allows the database to return some warnings along with
the response and the driver prints them using the tracing::warn! macro.
Those warnings, when reported by tracing, may be confused with a
driver-side error, so this commit attempts to make it more clear that
the warning was generated on the db-side.