transport: locator: add ReplicaLocator struct
`ReplicaLocator` provides a way to find the set of owning nodes for a
given (token, replication strategy) pair. It does so by either using the
precomputed token ranges, or doing the computation on the fly.

`ReplicaLocator::replicas_for_token` is the function that returns
replica sets (`ReplicaSet`). I chose to represent replica sets as a
separate type, to allow allocation-free construction & iteration through
precomputed replica sets.