caching_session: hide internals
The CachingSession should not expose its fields as public. It's not a
simple data object and its `max_capacity` and `cache` fields should not
be operated on directly.

This commit marks its fields as private and provides getters for
`max_capacity` and `session`. This should make it easier to modify the
internals of the CachingSession in the future without introducing
breaking changes to its interface.