caching_session: always inherit params from provided Query object
Currently, the CachingSession keeps full PreparedStatement objects in
its cache. When calling CachingSession::execute etc., a full Query
object is provided. If the query is not in the cache, then
Session::prepare is called and the resulting query - which inherits
parameters like paging size, consistency etc. - is inserted into the
cache. If the query is already in the cache, then the original
PreparedStatement which was inserted into the cache is returned - and
that prepared statement may have different parameters than the Query
passed as a parameter.

This commit modifies the CachingSession so that it only keeps essential
information necessary to create a PreparedStatement from a Query, i.e.
ID of the query and prepared statement metadata. All other parameters
are inherited from the Query provided as a parameter to the function.