caching_session: clone slightly bit less
The CachingSession::add_prepared_statement is now always forced to clone
most of the provided Query, which is passed by reference. It makes more
sense to pass the Query by value so that sometimes cloning can be
avoided. This actually helps in `execute` function and its friends
because they already take the Query by value.

Because `add_prepared_statement` is already public I decided to leave
the existing interface and add the `add_prepared_statement_owned` method
which is non-public and is used by `execute` methods.